backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
python-pptx
openpyxl
pdfplumber
anthropic
pydantic-settings
